validation works correctly as well as roles checkboxes
